includefoot (or inculdeall) should be used --
It's not very clear whether the standard allows to put the page
numbers (and other stuff) on the bottom margin, but it's probably safer not to do so
(so that the minimal bottom margin size is not violated in any case).

TODO: drop any stuff that occupies space in the head and use includeall.